FlashCardDisplayer component to hold card side state
