+++ lib/main.dart
@@ -6,6 +6,8 @@ import './models/FlashCard.dart';
 import './models/Deck.dart';
 import './models/Grade.dart';
 
+import './models/FlashCardDisplayer.dart';
+
 import './CardEditor.dart';
 import './DeckSelector.dart';
 import './ImportFromURL.dart';
@@ -42,7 +44,6 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  FlashCardSide _shownSide = FlashCardSide.front;
 
   List<FlashCard> _cardList;
   int _activeCardIdx = 0;
@@ -50,8 +51,6 @@ class _MyHomePageState extends State<MyHomePage> {
   int _cardGrade = -1;
   Deck _chosenDeck;
   PageController controller;
-
-  static final FlashCardSide BackSideOfCard = FlashCardSide.back;
 
   @override
   initState() {
@@ -74,71 +73,6 @@ class _MyHomePageState extends State<MyHomePage> {
     });
   }
 
-  void _flipCard() {
-    setState(() {
-      if (_shownSide == FlashCardSide.front) {
-        _shownSide = FlashCardSide.back;
-      } else {
-        _shownSide = FlashCardSide.front;
-      }
-    });
-  }
-
-  FractionallySizedBox cardToWidget(
-      {BuildContext context, FlashCardSide sideToDisplay, FlashCard card}) {
-    EdgeInsets contentPadding = EdgeInsets.all(24.0);
-    // Pad out the top of the card more if we are displying the rating row at bottom.
-    if (sideToDisplay == BackSideOfCard) {
-      contentPadding = EdgeInsets.fromLTRB(24.0, 48.0, 24.0, 0);
-    }
-    List<Widget> childList = [
-      Expanded(
-          flex: 1,
-          child: Center(
-            child: SingleChildScrollView(
-                padding: contentPadding,
-                child: Text(
-                  sideToDisplay == FlashCardSide.front ? card.front : card.back,
-                  style: TextStyle(fontWeight: FontWeight.w500, fontSize: 32),
-                  textAlign: TextAlign.center,
-                )),
-          ))
-    ];
-    if (sideToDisplay == BackSideOfCard) {
-      childList.add(Wrap(
-          spacing: 8.0,
-          children: List<Widget>.generate(
-            6,
-            (int index) {
-              return ChoiceChip(
-                backgroundColor: Colors.cyan.shade100,
-                selectedColor: Colors.orange,
-                // shape: StadiumBorder(side: BorderSide()),
-                selected: index == _cardGrade,
-                label: Text("$index",
-                    style: TextStyle(fontSize: 20, color: Colors.black)),
-                onSelected: (bool selected) {
-                  setState(() {
-                    _cardGrade = selected ? index : null;
-                  });
-                },
-              );
-            },
-          ).toList()));
-    }
-    return FractionallySizedBox(
-      widthFactor: 0.9,
-      heightFactor: 0.9,
-      child: Card(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: childList,
-        ),
-      ),
-    );
-  }
-
   Card buildNiceTextBox(textColumn) => Card(
         child: Column(
           mainAxisSize: MainAxisSize.min,
@@ -152,13 +86,7 @@ class _MyHomePageState extends State<MyHomePage> {
     if (_cardList != null &&
         _activeCardIdx >= 0 &&
         _activeCardIdx < _cardList.length) {
-      return GestureDetector(
-        child: cardToWidget(
-            context: context,
-            sideToDisplay: _shownSide,
-            card: _cardList[index]),
-        onTap: () => _flipCard(),
-      );
+      return FlashCardDisplayer(flashCard: _cardList[index]);
     } else {
       Widget textColumn = Column(children: <Widget>[
         Text("No cards available.",
@@ -253,7 +181,6 @@ class _MyHomePageState extends State<MyHomePage> {
             setState(() {
               _activeCardIdx = pageIdx;
               _cardGrade = -1;
-              _shownSide = FlashCardSide.front;
             });
           },
         ),
